refactor: meaningful const variables and add comments
