+++ src/utils/strategies/metricStrategy.js
@@ -1,7 +1,8 @@
-import validation from '../config/validation.json' with { type: 'json' };
+import validation from '../../config/validation.json' with { type: 'json' };
 
 export class MetricStrategy {
     calculateBMI(height, weight) {
+        // BMI formula requires height in meters; height is provided in centimeters, so divide by 100.
         const heightInMeters = height / 100;
         return weight / (heightInMeters ** 2);
     }
@@ -17,9 +18,10 @@ export class MetricStrategy {
     }
 
     getIdealWeightRange(height) {
-        const minBMI = 18.5;
-        const maxBMI = 24.9;
-        const heightInMeters = height / 100;
+        // 18.5 and 24.9 are chosen because they are the World Health Organization's recommended BMI thresholds for healthy weight.
+        const minBMI = 18.5; // lower healthy BMI limit
+        const maxBMI = 24.9; // upper healthy BMI limit
+        const heightInMeters = height / 100; // convert cm to meters
 
         return [
             minBMI * heightInMeters * heightInMeters,
